Move findAvailablePort to utils
